Conversation api added

+++ utils/initDb.js
@@ -9,7 +9,7 @@ const initDb = async (prismaConn) => {
     "content_title" TEXT,
     "content_url" TEXT,
     "content" TEXT,
-    "content_tokens" BIGINT,
+    "content_tokens" integer,
     "url_id" TEXT,
     "project_id" TEXT NOT NULL,
     "embedding" vector (1536),
@@ -65,7 +65,7 @@ const initDb = async (prismaConn) => {
   await prisma.$queryRaw`ALTER TABLE "conversations" ADD CONSTRAINT "conversations_project_id_fkey" FOREIGN KEY ("project_id") REFERENCES "projects"("project_id") ON DELETE NO ACTION ON UPDATE NO ACTION;`;
   await prisma.$queryRaw`ALTER TABLE "taskqueue" ADD CONSTRAINT "taskqueue_project_id_fkey" FOREIGN KEY ("project_id") REFERENCES "projects"("project_id") ON DELETE NO ACTION ON UPDATE NO ACTION;`;
   await prisma.$queryRaw`CREATE OR REPLACE FUNCTION public.embeddings_search(project_id text, query_embedding vector, similarity_threshold double precision, match_count integer)
-    RETURNS TABLE(id bigint, content_title text, content_url text, content text, content_tokens bigint, similarity double precision)
+    RETURNS TABLE(id integer, content_title text, content_url text, content text, content_tokens integer, similarity double precision)
     LANGUAGE plpgsql
    AS $function$
    BEGIN
